+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased](https://github.com/usgs/waterdataui/compare/waterdataui-0.25.0...master)
 ### Added
+- Daily value graph. This has a feature toggle and is valid for only a single site. Sites without
+data will show an info alert.
+### Added
 - Networks page.
 
 ## [0.25.0](https://github.com/usgs/waterdataui/compare/waterdataui-0.24.0...waterdataui-0.25.0)